requirements.txt
pyyaml
numpy
matplotlib
Pillow
pytest
pyquaternion
torch==1.5.0+cu101
torchvision==0.6.0+cu101
